requirements.txt
# Core web framework
fastapi==0.120.0
uvicorn==0.38.0
pydantic==2.12.3
python-multipart==0.0.9
python-dotenv==1.2.1
httpx>=0.27.0

# LangChain ecosystem
langchain==1.0.2
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-community==0.4.0
langchain-chroma==1.0.0
langchain-text-splitters==1.0.0
langchain-classic>=1.0.0

# Document and RAG processing
PyMuPDF==1.24.14
python-docx==1.1.2
docx2txt>=0.8
unstructured>=0.16.0

# OCR and image processing
easyocr==1.7.2
Pillow>=10.0.0
numpy>=2.0.0
opencv-python-headless>=4.8.0

# NLP and translation
nltk==3.9.2
langdetect==1.0.9
langid==1.1.6
deep-translator==1.11.4
regex>=2024.4.28

# Data analysis and visualization
pandas>=2.0.0
plotly==6.3.1
pyvis==0.3.2

# Industry intelligence and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
tls-client>=1.0.1
markdownify>=1.1.0

# Misc tools
wikipedia==1.4.0

# Optional dev/test tools
pytest==8.4.2
